tree_morph

Module cache

Define Caching behaviour for TreeMorph This should help out with repetetive and expensive tree operations as well as long chains of rules on duplicate subtrees

Structs

CachedHashKey

Hashing strategy that delegates to CacheHashable::get_cached_hash, allowing types with internally cached hashes to avoid rehashing entire subtrees.

HashMapCache

RewriteCache implemented with a HashMap, generic over a CacheKey hashing strategy.

NoCache

Disable Caching.

StdHashKey

Hashing strategy that delegates to the standard Hash trait.

Enums

CacheResult

Return type for RewriteCache Due to the nature of Rewriting, there may be repeated subtrees where no rule can be applied. In that case, we can compute it once and store it in cache stating no rules applicable.

Traits

CacheHashable

Types with an internally cached hash value.

CacheKey

Abstracts how a cache computes hash keys and invalidates nodes.

RewriteCache

Caching for Treemorph.

Type Aliases

CachedHashMapCache

Convenience alias for a HashMapCache using CachedHashKey.